requirements.txt
python-docx
zhipuai>=2.0.1
requests
sniffio
httpx
typing_extensions